feat: move GenSingleton processes to separate supervisor with full state reset on exit
